feat: easier ramping condition
Previous versions of cryptoswap had an eoa controlling the ramping so a more conservative condition was used to prevent multiple ramps in short time frames. Now that pools are controlled by the DAO this becomes much harder since consecutive ramps would require a chain of votes so we can simplify the ramping condition.