refactor: D is computed before `tweak_price`
`tweak_price` has been refactored to only compute `newton_D` when necessary for rebalance.

Previously it would also compute it if called by the exchange method which hurts the readability of the function.

Thanks to this change the argument `new_D` has now been renamed to `D_before_rebalance` for clarity and `K0_prev` has been removed since it's now directly used in the `–exchange` function, that was the only function overriding the default argument.