fix: donations do not inflate D
This commit fixes the faulty logic in donation which would make D (users liquidity) increase following a donation.

Since the pool now supports an operation that does not change the virtual price, the requirement for a strictly increasing virtual price has been modified to allow for an increasing or constant virtual price.