perf: reduce codesize by refactoring admin checks
A lot of space was taken by the revert string `"only owner"` strings since it was repeated every time a method checks for admin rights. Refactoring it into an helper method makes sure that it only appears once in the codebase. Without this optimisation it was basically impossible to add any methods to the contract without hitting EIP170 limits.